+++ CHANGELOG.md
@@ -6,10 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-## [1.7.0] - 2020-05-01
+
+## [1.7.0] - 2020-04-28
 ### Added
-- [#422](https://gitlab.com/dctx/rapidpass/rapidpass-api/-/issues/422) Add support for searching applicants 
-by mobile number 
+- [#403](https://gitlab.com/dctx/rapidpass/rapidpass-api/-/issues/403) Added Resend Text & Email Endpoint
+- [#425](https://gitlab.com/dctx/rapidpass/rapidpass-api/-/issues/425) Added `notified` property when querying a RapidPass.
 
 ## [1.6.4] - 2020-04-28
 ### Fixed